+++ src/main/java/com/springcooler/sgma/studygroup/query/service/StudyGroupService.java
@@ -1,47 +1,25 @@
 package com.springcooler.sgma.studygroup.query.service;
 
 import com.springcooler.sgma.studygroup.query.dto.StudyGroupDTO;
-import com.springcooler.sgma.studygroup.query.repository.StudyGroupMapper;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-@Service
-public class StudyGroupService {
-    private final StudyGroupMapper studyGroupMapper;
-
-    @Autowired
-    public StudyGroupService(StudyGroupMapper studyGroupMapper) {
-        this.studyGroupMapper = studyGroupMapper;
-    }
-
+public interface StudyGroupService {
     // 스터디 그룹 전체 조회
-    public List<StudyGroupDTO> findAllStudyGroups() {
-        return studyGroupMapper.findAllStudyGroups();
-    }
+    List<StudyGroupDTO> findAllStudyGroups();
 
     // 생성한 스터디 그룹 조회
-    public List<StudyGroupDTO> findStudyGroupsByOwnerId(Long ownerId) {
-        return studyGroupMapper.findStudyGroupsByOwnerId(ownerId);
-    }
+    List<StudyGroupDTO> findStudyGroupsByOwnerId(Long ownerId);
 
     // 참가한 스터디 그룹 조회
-    public List<StudyGroupDTO> findStudyGroupsByParticipantId(Long participantId) {
-        return studyGroupMapper.findStudyGroupsByParticipantId(participantId);
-    }
+    List<StudyGroupDTO> findStudyGroupsByParticipantId(Long participantId);
 
     // 스터디 그룹 카테고리별 조회
-    public List<StudyGroupDTO> findStudyGroupsByCategoryId(int categoryId) {
-        return studyGroupMapper.findStudyGroupsByCategoryId(categoryId);
-    }
+    List<StudyGroupDTO> findStudyGroupsByCategoryId(int categoryId);
 
-    // 스터디 그룹 단건 조회
-    public List<StudyGroupDTO> findStudyGroupByGroupId(long groupId) {
-        return studyGroupMapper.findStudyGroupByGroupId(groupId);
-    }
+    // 스터디 그룹 단건 조회(그룹 아이디)
+    List<StudyGroupDTO> findStudyGroupByGroupId(long groupId);
 
-    public List<StudyGroupDTO> findStudyGroupByGroupName(String groupName) {
-        return studyGroupMapper.findStudyGroupByGroupName(groupName);
-    }
+    // 스터디 그룹 단건 조회(그룹 이름)
+    List<StudyGroupDTO> findStudyGroupByGroupName(String groupName);
 }
